refactor: consolidate settings unit tests
